refactor: url 상태관리를 커스텀 훅에서 수행하도록 변경
